Fixing pyb11 DataBase warnings.

+++ src/DataBase/DataBaseInline.hh
@@ -418,7 +418,7 @@ resizeSolidFieldList(FieldList<Dimension, DataType>& fieldList,
   VERIFY((fieldList.storageType() == FieldStorageType::CopyFields));
 
   // First check if it's necessary to resize the FieldList.
-  bool reinitialize = fieldList.numFields() != numSolidNodeLists();
+  bool reinitialize = (int)fieldList.numFields() != numSolidNodeLists();
   ConstSolidNodeListIterator nodeListItr = solidNodeListBegin();
   typename FieldList<Dimension, DataType>::const_iterator itr = fieldList.begin();
   while (!reinitialize &&
